requirements.txt
pandas
scipy
plotly
requests
scikit-learn
openpyxl
python-dotenv